Spielerzustand aktualisieren
