+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/game/GameStateClient.kt
@@ -1,14 +1,5 @@
 package at.aau.serg.websocketbrokerdemo.game
 
-data class PlayerState(
-    val id: String,
-    var name: String,
-    var position: Int,
-    var money: Int,
-    var properties: List<Int>,  // IDs der gekauften Tiles
-    var inJail: Boolean
-)
-
 object GameStateClient {
     private val playerPositions = mutableMapOf<String, Int>()
     private val players: MutableMap<String, PlayerState> = mutableMapOf()
